# packages/ts-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cocoon-SIT API",
    "description": "Cocoon-SIT backend implementing REST 202 + WebSocket realtime contracts, versioned prompt templates, audit artifacts, and worker-driven runtime execution.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/providers": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "List Providers",
        "operationId": "list_providers_api_v1_providers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelProviderOut"
                  },
                  "type": "array",
                  "title": "Response List Providers Api V1 Providers Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Create Provider",
        "operationId": "create_provider_api_v1_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/providers/{provider_id}": {
      "patch": {
        "tags": [
          "providers"
        ],
        "summary": "Update Provider",
        "operationId": "update_provider_api_v1_providers__provider_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelProviderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "providers"
        ],
        "summary": "Delete Provider",
        "operationId": "delete_provider_api_v1_providers__provider_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelProviderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{provider_id}/sync-models": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Sync Provider Models",
        "operationId": "sync_provider_models_api_v1_providers__provider_id__sync_models_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailableModelOut"
                  },
                  "title": "Response Sync Provider Models Api V1 Providers  Provider Id  Sync Models Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/providers/{provider_id}/test": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Test Provider",
        "operationId": "test_provider_api_v1_providers__provider_id__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderTestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tags": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "List Tags",
        "operationId": "list_tags_api_v1_tags_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagOut"
                  },
                  "type": "array",
                  "title": "Response List Tags Api V1 Tags Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "tags"
        ],
        "summary": "Create Tag",
        "operationId": "create_tag_api_v1_tags_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/tags/{tag_id}": {
      "patch": {
        "tags": [
          "tags"
        ],
        "summary": "Update Tag",
        "operationId": "update_tag_api_v1_tags__tag_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tags"
        ],
        "summary": "Delete Tag",
        "operationId": "delete_tag_api_v1_tags__tag_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get System Settings",
        "operationId": "get_system_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update System Settings",
        "operationId": "update_system_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/cocoons/{cocoon_id}/tags": {
      "get": {
        "tags": [
          "cocoons"
        ],
        "summary": "List Cocoon Tags",
        "operationId": "list_cocoon_tags_api_v1_cocoons__cocoon_id__tags_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CocoonTagBindingOut"
                  },
                  "title": "Response List Cocoon Tags Api V1 Cocoons  Cocoon Id  Tags Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cocoons"
        ],
        "summary": "Bind Cocoon Tag",
        "operationId": "bind_cocoon_tag_api_v1_cocoons__cocoon_id__tags_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CocoonTagBindRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CocoonTagBindResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cocoons/{cocoon_id}/tags/{tag_id}": {
      "delete": {
        "tags": [
          "cocoons"
        ],
        "summary": "Unbind Cocoon Tag",
        "operationId": "unbind_cocoon_tag_api_v1_cocoons__cocoon_id__tags__tag_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CocoonTagBindResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailableModelOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_kind": {
            "type": "string",
            "title": "Model Kind"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default"
          },
          "config_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider_id",
          "model_name",
          "model_kind",
          "is_default",
          "config_json",
          "created_at"
        ],
        "title": "AvailableModelOut"
      },
...
      "CocoonTagBindRequest": {
        "properties": {
          "tag_id": {
            "type": "string",
            "title": "Tag Id"
          }
        },
        "type": "object",
        "required": [
          "tag_id"
        ],
        "title": "CocoonTagBindRequest"
      },
      "CocoonTagBindResult": {
        "properties": {
          "binding_id": {
            "type": "string",
            "title": "Binding Id"
          },
          "tag_id": {
            "type": "string",
            "title": "Tag Id"
          }
        },
        "type": "object",
        "required": [
          "binding_id",
          "tag_id"
        ],
        "title": "CocoonTagBindResult"
      },
      "CocoonTagBindingOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tag_id": {
            "type": "string",
            "title": "Tag Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tag_id",
          "created_at"
        ],
        "title": "CocoonTagBindingOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelProviderCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "openai_compatible"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "capabilities_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities Json"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ModelProviderCreate"
      },
      "ModelProviderOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled"
          },
          "capabilities_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "base_url",
          "is_enabled",
          "capabilities_json",
          "created_at"
        ],
        "title": "ModelProviderOut"
      },
...
      "ProviderTestOut": {
        "properties": {
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "selected_model_id": {
            "type": "string",
            "title": "Selected Model Id"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "usage": {
            "$ref": "#/components/schemas/ProviderUsageOut"
          },
          "raw_response": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Response"
          }
        },
        "type": "object",
        "required": [
          "provider_id",
          "selected_model_id",
          "model_name",
          "reply",
          "usage",
          "raw_response"
        ],
        "title": "ProviderTestOut"
      },
      "ProviderTestRequest": {
        "properties": {
          "selected_model_id": {
            "type": "string",
            "title": "Selected Model Id"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": [
          "selected_model_id",
          "prompt"
        ],
        "title": "ProviderTestRequest"
      },
      "ProviderUsageOut": {
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "title": "Prompt Tokens",
            "default": 0
          },
          "completion_tokens": {
            "type": "integer",
            "title": "Completion Tokens",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "default": 0
          }
        },
        "type": "object",
        "title": "ProviderUsageOut"
      },
...
      "SystemSettingsOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "allow_registration": {
            "type": "boolean",
            "title": "Allow Registration"
          },
          "max_chat_turns": {
            "type": "integer",
            "title": "Max Chat Turns"
          },
          "allowed_model_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Model Ids"
          },
          "default_cocoon_temperature": {
            "type": "number",
            "title": "Default Cocoon Temperature"
          },
          "default_max_context_messages": {
            "type": "integer",
            "title": "Default Max Context Messages"
          },
          "default_auto_compaction_enabled": {
            "type": "boolean",
            "title": "Default Auto Compaction Enabled"
          },
          "private_chat_debounce_seconds": {
            "type": "integer",
            "title": "Private Chat Debounce Seconds"
          },
          "rollback_retention_days": {
            "type": "integer",
            "title": "Rollback Retention Days"
          },
          "rollback_cleanup_interval_hours": {
            "type": "integer",
            "title": "Rollback Cleanup Interval Hours"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "allow_registration",
          "max_chat_turns",
          "allowed_model_ids",
          "default_cocoon_temperature",
          "default_max_context_messages",
          "default_auto_compaction_enabled",
          "private_chat_debounce_seconds",
          "rollback_retention_days",
          "rollback_cleanup_interval_hours",
          "created_at",
          "updated_at"
        ],
        "title": "SystemSettingsOut"
      },
...
      "TagCreate": {
        "properties": {
          "tag_id": {
            "type": "string",
            "minLength": 1,
            "title": "Tag Id"
          },
          "brief": {
            "type": "string",
            "minLength": 1,
            "title": "Brief"
          },
          "is_isolated": {
            "type": "boolean",
            "title": "Is Isolated",
            "default": false
          },
          "meta_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta Json"
          }
        },
        "type": "object",
        "required": [
          "tag_id",
          "brief"
        ],
        "title": "TagCreate"
      },
      "TagOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tag_id": {
            "type": "string",
            "title": "Tag Id"
          },
          "brief": {
            "type": "string",
            "title": "Brief"
          },
          "is_isolated": {
            "type": "boolean",
            "title": "Is Isolated"
          },
          "meta_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tag_id",
          "brief",
          "is_isolated",
          "meta_json",
          "created_at"
        ],
        "title": "TagOut"
      },
      "TagUpdate": {
        "properties": {
          "brief": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brief"
          },
          "is_isolated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Isolated"
          },
          "meta_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Json"
          }
        },
        "type": "object",
        "title": "TagUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}